Move classes into separate files, add default spec task to Rakefile
