+++ Rakefile
@@ -1,7 +1,12 @@
-require 'rubygems'
-require 'bundler'
-Bundler::GemHelper.install_tasks
+require "bundler/gem_tasks"
 
-require 'rspec/core/rake_task'
-rspec = RSpec::Core::RakeTask.new
-rspec.rspec_opts = '-f Fuubar --color'
+
+begin
+  require 'rspec/core/rake_task'
+
+  RSpec::Core::RakeTask.new(:spec)
+
+  task :default => :spec
+rescue LoadError
+  # no rspec available
+end
