+++ spec/spec_helper.rb
@@ -1,0 +1,4 @@
+require 'rspec'
+
+require 'campfire_export'
+require 'tzinfo'
